refactor: ValidateAuthenticationHeader is now static
